+++ react/__mocks__/vtex.checkout-resources/QueryOrderForm.ts
@@ -1,11 +1,16 @@
 import gql from 'graphql-tag'
 
 const orderForm = gql`
-  query MockQuery {
-    orderForm {
+  query MockQuery($refreshOutdatedData: Boolean) {
+    orderForm(refreshOutdatedData: $refreshOutdatedData) {
       id
       items
       canEditData
+      paymentData {
+        installmentOptions {
+          value
+        }
+      }
       clientProfileData {
         email
         firstName