+++ react/__mocks__/vtex.render-runtime.ts
@@ -1,1 +1,8 @@
 export const useSSR = () => false
+
+export const mockedRuntimeHook = (page = '') => ({
+  rootPath: '',
+  page,
+})
+
+export const useRuntime = jest.fn(mockedRuntimeHook)