+++ react/__tests__/OrderForm.test.tsx
@@ -1,26 +1,37 @@
 import type { FunctionComponent } from 'react'
 import React, { useEffect, useCallback } from 'react'
-import {
-  fireEvent,
-  render,
-  wait,
-  act,
-  flushPromises,
-} from '@vtex/test-tools/react'
+import { fireEvent, render, act, flushPromises } from '@vtex/test-tools/react'
 import type { Item } from 'vtex.checkout-graphql'
 import OrderForm from 'vtex.checkout-resources/QueryOrderForm'
+import * as renderRuntime from 'vtex.render-runtime'
 
-import { mockOrderForm } from '../__fixtures__/orderForm'
+import {
+  mockOrderForm,
+  refreshedMockOrderForm,
+} from '../__fixtures__/orderForm'
 import { OrderFormProvider, useOrderForm } from '../OrderForm'
 import { OrderQueueProvider } from '../OrderQueue'
 
 const mockQuery = {
   request: {
     query: OrderForm,
+    variables: { refreshOutdatedData: false },
   },
   result: {
     data: {
       orderForm: mockOrderForm,
+    },
+  },
+}
+
+const refreshedMockQuery = {
+  request: {
+    query: OrderForm,
+    variables: { refreshOutdatedData: true },
+  },
+  result: {
+    data: {
+      orderForm: refreshedMockOrderForm,
     },
   },
 }
@@ -32,6 +43,70 @@ describe('OrderForm', () => {
 
   afterEach(() => {
     localStorage.clear()
+  })
+
+  it('should refresh outdated data on entering checkout', async () => {
+    const mockedUseRuntime = jest
+      .spyOn(renderRuntime, 'useRuntime')
+      .mockImplementation(
+        // @ts-expect-error: we do not want to mock the whole
+        // runtime object, only the page
+        () => ({ page: 'product' })
+      )
+
+    const Component: FunctionComponent = () => {
+      const { orderForm } = useOrderForm()
+
+      return (
+        <div>
+          Installment: {orderForm.paymentData?.installmentOptions?.[0]?.value}
+        </div>
+      )
+    }
+
+    const { queryByText, rerender } = render(
+      <OrderQueueProvider>
+        <OrderFormProvider>
+          <Component />
+        </OrderFormProvider>
+      </OrderQueueProvider>,
+      { graphql: { mocks: [mockQuery, refreshedMockQuery] } }
+    )
+
+    act(() => jest.runAllTimers())
+
+    await act(async () => {
+      await new Promise<void>((resolve) => resolve())
+    })
+
+    const installmentParagraph = queryByText(/installment: /i)
+
+    expect(installmentParagraph).toHaveTextContent(/installment: 100/i)
+
+    mockedUseRuntime.mockImplementation(
+      // @ts-expect-error: same as above
+      () => ({ page: 'checkout' })
+    )
+
+    rerender(
+      <OrderQueueProvider>
+        <OrderFormProvider>
+          <Component />
+        </OrderFormProvider>
+      </OrderQueueProvider>
+    )
+
+    act(() => jest.runAllTimers())
+
+    await act(async () => {
+      await new Promise<void>((resolve) => resolve())
+    })
+
+    expect(installmentParagraph).toHaveTextContent(/installment: 200/i)
+
+    mockedUseRuntime.mockImplementation(
+      jest.requireMock('vtex.render-runtime').mockedRuntimeHook
+    )
   })
 
   it('should be possible to update order form with an update function', async () => {
@@ -143,9 +218,9 @@ describe('OrderForm', () => {
       await flushPromises()
     })
 
-    expect(getByText(mockOrderForm.items[0].name)).toBeTruthy()
-    expect(getByText(mockOrderForm.items[1].name)).toBeTruthy()
-    expect(getByText(mockOrderForm.items[2].name)).toBeTruthy()
+    expect(getByText(mockOrderForm.items[0].name ?? '')).toBeTruthy()
+    expect(getByText(mockOrderForm.items[1].name ?? '')).toBeTruthy()
+    expect(getByText(mockOrderForm.items[2].name ?? '')).toBeTruthy()
   })
 
   it('should correctly update the order form', async () => {
@@ -197,6 +272,7 @@ describe('OrderForm', () => {
       const orderFormMockQuery = {
         request: {
           query: OrderForm,
+          variables: { refreshOutdatedData: false },
         },
         result: {
           data: {
@@ -253,6 +329,7 @@ describe('OrderForm', () => {
       const orderFormMockQuery = {
         request: {
           query: OrderForm,
+          variables: { refreshOutdatedData: false },
         },
         result: {
           data: {